:bug: Change Rate type to Double (#370)
* :bug: Change Rate type to Float

* :construction: Use doubles instead of floats

+++ Toggl.Foundation/Models/FoundationModels.cs
@@ -50,7 +50,7 @@ namespace Toggl.Foundation.Models
 
         public int? EstimatedHours { get; }
 
-        public int? Rate { get; }
+        public double? Rate { get; }
 
         public string Currency { get; }
